Merged from github fork.

# By Brett Wooldridge (40) and brettwooldridge (11)
* commit 'df394d462ca14412ab724a4e5dcb4ecf75ac60dc': (51 commits)
  Cleanup compiler warnings, mostly generics.
  Cleanup compiler warnings, mostly generics.
  Reorganize the order of methods in XAPool for increased readability.  Pull static XAFactory creation out into a helper class (XAFactoryHelper).
  Don't allow attempt to increase pool to interfere with returning a connection that has tested valid.
  Fix unconditional return
  Ensure the pool shrink task runs even if maxIdleTime isn't configured but when maxLifeTime is.   Remove extra sleep on connection failure.
  Clean up/tighten up the Javassist proxy generation code.
  Remove some unused code.
  A little sanity check logging.
  CallableStatement proxy was being created with the PreparedStatement interface.
  Use AtomicReference to hold active transaction log
  Fix possible boundary cases.
  Fix class loader visibility issue when in OSGi container (or maybe other app server containers).
  Add Javassist-based proxy generation support.
  Use the factory.
  Reuse the same JdbcJavaProxyFactory for creating LRC proxies.
  Methods and members don't need to be static here.
  Switch to caching the proxy class contractors for improved performance instantiating Java Proxy-based proxies.
  If an interface is passed into getAllInterfaces() add that interface itself to the returned set.
  Add file header
  ...

Conflicts:
	btm-jdbc3-stubs/pom.xml
	btm-jdbc4-stubs/pom.xml
	btm/src/main/java/bitronix/tm/Configuration.java
	btm/src/main/java/bitronix/tm/journal/DiskJournal.java
	btm/src/main/java/bitronix/tm/journal/TransactionLogAppender.java
	btm/src/main/java/bitronix/tm/journal/TransactionLogHeader.java
	btm/src/main/java/bitronix/tm/journal/TransactionLogRecord.java
	btm/src/main/java/bitronix/tm/resource/common/XAPool.java
	btm/src/main/java/bitronix/tm/resource/jdbc/JdbcUncachedPreparedStatementHandle.java
	btm/src/test/java/bitronix/tm/journal/DiskJournalTest.java

+++ btm/src/main/java/bitronix/tm/resource/jdbc/PoolingDataSourceMBean.java
@@ -28,8 +28,8 @@ public interface PoolingDataSourceMBean {
 
     public int getMinPoolSize();
     public int getMaxPoolSize();
-    public long getInPoolSize();
-    public long getTotalPoolSize();
+    public int getInPoolSize();
+    public int getTotalPoolSize();
     public boolean isFailed();
     public void reset() throws Exception;
     public boolean isDisabled();
